+++ spec/error_responses/jsonapi_spec.rb
@@ -17,6 +17,34 @@ RSpec.describe OpenapiFirst::ErrorResponses::Jsonapi do
     expect(response.content_type).to eq 'application/vnd.api+json'
   end
 
+  context 'with duplicated errors' do
+    subject(:error_response) do
+      error = instance_double(OpenapiFirst::Schema::ValidationError, message: 'hey', data_pointer: '/data', type: 'fail')
+      described_class.new(
+        failure: OpenapiFirst::Failure.new(
+          :invalid_body,
+          errors: [error, error]
+        )
+      )
+    end
+
+    it 'returns unique errors' do
+      response = Rack::MockResponse[*error_response.render]
+      expect(response.status).to eq(400)
+      expect(JSON.parse(response.body, symbolize_names: true).fetch(:errors)).to eq(
+        [
+          {
+            title: 'hey',
+            source: { pointer: '/data' },
+            code: 'fail',
+            status: '400'
+
+          }
+        ]
+      )
+    end
+  end
+
   context 'with invalid body' do
     let(:schema) do
       {